% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BinTest}
\alias{BinTest}
\title{A test function provide binomial data for testing Weierstrass rejection sampling (and comparing to the fullset posterior/averaging/weighted averaging combiners).}
\usage{
BinTest(prob = 0.001, samp = 20000, n = 20000, m = 20, accept = 0.1,
  draw = TRUE)
}
\arguments{
\item{prob}{The parameter for binomial distribution. Default is 0.001.}

\item{samp}{The posterior sample size on each subset. Default is 20000.}

\item{n}{The total sample size. Default is 10000.}

\item{m}{Number of subsets. Default is 20.}

\item{accept}{The acceptance rate for the Weierstrass rejection sampling. Default is 0.2.}

\item{draw}{Indicate whether the result should be plotted.}
}
\value{
A list containing several components.
\enumerate{
\item posteriors: a list containing all subset posterior samples. Input for weierstrass rejection samping.
\item true.posterior: a matrix containing posterior samples drawn with full data set.
\item CombSample.weight: a matrix containing combined samples generated by the Weierstrass rejection sampler
\item CombSample.unweight: a matrix containing combined samples generated by the unweighted Weierstrass rejection sampler
\item weight.ave: a matrix containing combined samples via inverse-variance weighted averaging.
\item ave: a matrix containing combined samples via simple averaging.
}
}
\description{
A test function provide binomial data for testing Weierstrass rejection sampling (and comparing to the fullset posterior/averaging/weighted averaging combiners).
}
\details{
The function generates data from a binomial distribution and make use of Weierstrass rejection sampler, weighted average and averge to combine the subset posterior samples.Both the weighted and unweighted weierstrass sampler will be used, and the results are saved for further processing.
}
\examples{
\dontrun{BinTest()}
\dontrun{BinTest(prob = 0.49)}
\dontrun{BinTest(prob = 0.99)}
}
\seealso{
\code{\link{weierstrass}} for the details of weierstrass rejection sampling. \code{\link{logitTest}} for another test on the logistic regression.
}

