% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logitTest}
\alias{logitTest}
\title{A test function provide logistic regression (require package "BayesLogit") for testing Weierstrass rejection sampling (and comparing to the fullset posterior/averaging/weighted averaging combiners).}
\usage{
logitTest(n = 20000, samp = 20000, p = 5, m = 20, r = 0.3,
  draw = TRUE, accept = 0.1)
}
\arguments{
\item{n}{The total sample size. Default is 10000.}

\item{samp}{The posterior sample size on each subset. Default is 20000.}

\item{p}{The number of predictors (intercept is always included, so the total number of coefficients will be p+1). Default is 5.}

\item{m}{Number of subsets. Default is 20.}

\item{r}{The correlation between predictors. Default is 0.3.}

\item{draw}{Indicate whether the result should be plotted.}

\item{accept}{The acceptance rate for the Weierstrass rejection sampling.}
}
\value{
A list containing several components.
\enumerate{
\item Samples: a list containing all subset posterior samples. Input for weierstrass rejection samping.
\item true.posterior: a matrix containing posterior samples drawn with full data set.
\item CombSample.weight: a matrix containing combined samples generated by the Weierstrass rejection sampler
\item CombSample.unweight: a matrix containing combined samples generated by the unweighted Weierstrass rejection sampler
\item weight.ave: a matrix containing combined samples via inverse-variance weighted averaging.
\item ave: a matrix containing combined samples via simple averaging.
}
}
\description{
A test function provide logistic regression (require package "BayesLogit") for testing Weierstrass rejection sampling (and comparing to the fullset posterior/averaging/weighted averaging combiners).
}
\details{
The function generates data from a logistic regression and make use of Weierstrass rejection sampler, weighted average and averge to combine the subset posterior samples. The coefficients of the model is generated by the following formula
\deqn{\beta_i \sim (-1)^{ber(0.6)}|N(0,4)|}
Both the weighted and unweighted weierstrass sampler will be used, and the results are saved for further processing.
}
\examples{
\dontrun{logitTest()}
\dontrun{logitTest(n = 20000, p = 10, m = 50, r = 0.4)}
}
\seealso{
\code{\link{weierstrass}} for the details of weierstrass rejection sampling. \code{\link{BinTest}} for another test on the binomial data.
}

